Refactor: service 시그니처 변경
- #10
