+++ src/main/java/co/wanted/board/api/member/application/SearchService.java
@@ -1,12 +1,16 @@
 package co.wanted.board.api.member.application;
 
+import co.wanted.board.api.member.application.exception.ErrorCode;
+import co.wanted.board.api.member.application.exception.MemberException;
+import co.wanted.board.api.member.domain.Member;
 import co.wanted.board.api.member.infrastructure.persistence.MemberRepository;
+import co.wanted.board.api.member.presentation.dto.Signup;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MemberService {
+public class SearchService {
 
     private final MemberRepository memberRepository;
 
@@ -17,4 +21,9 @@ public class MemberService {
     public boolean isExistName(String username) {
         return memberRepository.existsByUsername(username);
     }
+
+    public Member getMemberByEmail (String email) {
+        return memberRepository.findByEmail(email).orElseThrow(() -> new MemberException(ErrorCode.NO_SUCH_MEMBER));
+    }
+
 }
